Queues access request confirmation email
Adds functionality to queue an email to the user confirming
that their access request has been submitted. It configures
and adds the email queue table and adds a default status of
pending. The queueing is triggered upon a successful request.

Implements the email template, app name, and destination email
as environment variables in the config.

Relates to #16