Enhances user access management and security (#54)
* Enables updating user group assignments

Adds functionality to update a user's group assignment via the API.

This change introduces a new endpoint for updating user group memberships.
It also updates the user service and authentication service to handle the
group update logic and integrates with the Auth API.
The web UI is updated to allow editing a user's access level, using the new API endpoint.

Relates to #45

* Implements rate limiting for user initialization

Adds rate limiting to the user initialization endpoint to prevent abuse.

The rate limit is set to 5 requests per minute, using a user identifier derived from either the JWT token or the remote address.

Also includes:
- Adds `flask-limiter` to the project's dependencies.
- Includes drag and drop functionality to the web application.
- Adds retry logic to user settings API calls.

* Renames User model to avoid naming conflicts

Updates the User model import in the API to use an alias
(UserModel) to avoid potential naming conflicts with other User
related entities in the codebase. This change also removes an
unnecessary console log statement from the web application.

Related to CENTRE-task#45-C

* Moves limiter initialization to extensions

Moves the Limiter initialization to the extensions module
for better organization and to avoid circular dependencies.
This encapsulates the limiter configuration and enhances
modularity within the application.

Related to CENTRE-task#45-C

* Adds a docstring to the extensions module

Adds a docstring to the extensions module to provide a brief explanation of its purpose, which is to initialize and configure the Flask-Limiter extension for rate limiting.

Relates to CENTRE-task#45-C

* Adds docstring and uses consistent quotes

Adds a docstring to the `get_user_identifier` function for better documentation.

Uses consistent single quotes for string literals to maintain code style consistency.

Relates to CENTRE-task#45-C