Enables user-defined application sort order
Allows users to customize the order of applications displayed in the launchpad.

This change introduces a sort_order field to the UserApplication model and
implements a bulk upsert method for efficient updates.

A new API endpoint is added to handle PATCH requests for updating the sort
order, and the frontend is updated to allow drag-and-drop reordering of
applications.

The user initialization endpoint and model were removed in favour of
using user auth guid.

Relates to CENTRE-task#45